refactor(ui): migrate shared interaction primitives
