+++ src/renderer/src/components/anonymizer/label-manager/remove-dialog.tsx
@@ -1,10 +1,10 @@
-import Button from "@/components/ui/button";
 import {
+  Button,
   Dialog,
   DialogContent,
   DialogFooter,
   DialogTitle,
-} from "@/components/ui/dialog";
+} from "@aymurai/ui";
 
 interface RemoveDialogProps {
   isOpen: boolean;
