+++ src/renderer/src/components/file/suggestion-label.tsx
@@ -1,7 +1,7 @@
-import Suggestion from "@/components/ui/suggestion";
 import { css, cx } from "@/styled/css";
 import { HStack, styled } from "@/styled/jsx";
 import type { AllLabels, AllLabelsWithSufix } from "@/types/aymurai";
+import { Suggestion } from "@aymurai/ui";
 import { type ComponentPropsWithoutRef, forwardRef } from "react";
 
 interface SuggestionLabelProps
