+++ src/renderer/src/components/finish/finish-dataset.tsx
@@ -4,11 +4,11 @@ import { HStack } from "@/styled/jsx";
 import { FeatureFlowEnum } from "@/types/features";
 import type { DocFile } from "@/types/file";
 import { submitValidations } from "@/utils/file";
+import { Button } from "@aymurai/ui";
 import { useEffect, useState } from "react";
 import { useTranslation } from "react-i18next";
 import FileCheck from "../file-check";
 import Footer from "../layout/footer";
-import Button from "../ui/button";
 import FinishMainContent from "./finish-main-content";
 
 interface FinishDatasetProps {
