+++ src/renderer/src/components/validation-form/index.tsx
@@ -12,10 +12,10 @@ import {
 } from "react";
 
 import type { NativeComponent } from "@/types/component";
+import { Button } from "@aymurai/ui";
 
 import { css } from "@/styled/css";
 import { styled } from "@/styled/jsx";
-import Button from "../ui/button";
 import { Form } from "./ValidationForm.styles";
 
 interface Props extends NativeComponent<"form"> {
